fix(lens): consolidate xterm private _core reaches into one xtermInternals module
Move readBufferBytes (Terminal.tsx) and renderService/readDecPrivateMode (renderRecovery.ts) behind one null-guarded xtermInternals.ts, so the single volatility axis — xterm renaming a private _core field — has one receptacle instead of two.

Agreed by the lowy ⇄ hickey lens debate (finding lowy-1, raised by lowy). Not pushed or merged.
